populate db (buildings and animals)

+++ Backend/ZooBuilderBackend/ZooBuilderBackend/Models/Animal.cs
@@ -4,7 +4,7 @@ public class Animal
 {
     public int Id { get; set; }
     public string Species { get; set; }
-    public decimal Costs { get; set; }
+    public int Costs { get; set; }
     public int Hunger { get; set; }
     public string Diet { get; set; }
     public int Attraction { get; set; }
