+++ Backend/ZooBuilderBackend/ZooBuilderBackend/Models/Zoo.cs
@@ -4,7 +4,7 @@ public class Zoo
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public decimal Money { get; set; }
+    public int Money { get; set; }
     public int Vegetables { get; set; }
 
     public int PlayerId { get; set; }
